Move codecs away from persistance module
