+++ src/main/scala/mailo/persistence/EmailPersistanceActor.scala
@@ -6,28 +6,9 @@ import akka.persistence._
 import io.circe.syntax._
 import io.circe.generic.auto._
 
-import io.circe.{Decoder, Encoder, HCursor, Json}
-import akka.http.scaladsl.model.ContentType
-import io.circe.DecodingFailure
-
 case class CleanQueue(lastSequenceNr: Long)
 case class SendEmail(email: Mail)
 case class EmailEvent(content: String)
-
-trait CustomCodecs {
-  implicit val encodeFoo: Encoder[ContentType] = new Encoder[ContentType] {
-    final def apply(c: ContentType): Json =
-      Json.obj(("contentType", Json.fromString(c.value)))
-  }
-
-  implicit val decodeFoo: Decoder[ContentType] = new Decoder[ContentType] {
-    final def apply(c: HCursor): Decoder.Result[ContentType] =
-      for {
-        contentType <- c.downField("contentType").as[String]
-        result <- ContentType.parse(contentType).left.map(_ => DecodingFailure("Falied decoding content type", Nil))
-      } yield result
-  }
-}
 
 case class EmailState(events: List[EmailEvent] = Nil) {
   def updated(evt: EmailEvent): EmailState = copy(evt :: events)
@@ -36,7 +17,7 @@ case class EmailState(events: List[EmailEvent] = Nil) {
 }
 
 class EmailPersistorActor extends PersistentActor
-    with ActorLogging with CustomCodecs {
+    with ActorLogging with CustomContentTypeCodecs {
   override def persistenceId = "emails-persistence"
 
   private[this] var state: EmailState = EmailState()
